backend/ai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
psycopg2-binary==2.9.9
fastembed==0.4.1
openai>=1.0.0
msal==1.31.0
requests==2.32.3
tenacity==9.0.0
apscheduler==3.10.4
pydantic==2.9.0
python-dotenv==1.0.1
